# docs/swagger.yaml
basePath: /
definitions:
...
  dto.AlertResponse:
    properties:
      created_at:
        type: string
      created_by:
        type: integer
      description:
        type: string
      id:
        type: integer
      priority:
        $ref: '#/definitions/entities.AlertPriority'
      ref_id:
        type: integer
      ref_type:
        type: string
      resolved_at:
        type: string
      status:
        $ref: '#/definitions/entities.AlertStatus'
      title:
        type: string
      type:
        $ref: '#/definitions/entities.AlertType'
    type: object
...
  dto.CreateAlertRequest:
    properties:
      description:
        type: string
      priority:
        allOf:
        - $ref: '#/definitions/entities.AlertPriority'
        enum:
        - low
        - medium
        - high
      ref_id:
        type: integer
      ref_type:
        type: string
      title:
        minLength: 3
        type: string
      type:
        allOf:
        - $ref: '#/definitions/entities.AlertType'
        enum:
        - low_stock
        - expired_input
        - pest
        - weather
        - pending_application
        - harvest_soon
    required:
    - title
    - type
    type: object
...
  dto.CreateHarvestRequest:
    properties:
      field_loss:
        maximum: 100
        minimum: 0
        type: number
      grain_moisture:
        maximum: 100
        minimum: 0
        type: number
      harvest_date:
        type: string
      impurity:
        maximum: 100
        minimum: 0
        type: number
      notes:
        type: string
      planting_id:
        type: integer
      productivity_bag_ha:
        minimum: 0
        type: number
      productivity_kg_ha:
        minimum: 0
        type: number
      total_bags:
        minimum: 0
        type: number
    required:
    - harvest_date
    - planting_id
    type: object
...
  dto.CreateMonitoringRequest:
    properties:
      crop_stage:
        type: string
      field_id:
        type: integer
      infestation_level:
        maximum: 100
        minimum: 0
        type: number
      inspection_date:
        type: string
      inspector:
        type: string
      notes:
        type: string
      planting_id:
        type: integer
      problem_name:
        minLength: 2
        type: string
      severity:
        allOf:
        - $ref: '#/definitions/entities.Severity'
        enum:
        - low
        - moderate
        - high
        - critical
      technical_rec:
        type: string
      type:
        allOf:
        - $ref: '#/definitions/entities.MonitoringType'
        enum:
        - insect
        - disease
        - weed
        - nutritional
      urgent:
        type: boolean
    required:
    - field_id
    - inspection_date
    - inspector
    - problem_name
    - severity
    - type
    type: object
...
  dto.DashboardResponse:
    properties:
      active_plantings:
        type: integer
      applications_this_month:
        type: integer
      last_alerts:
        items:
          $ref: '#/definitions/dto.AlertResponse'
        type: array
      low_stock_inputs:
        type: integer
      open_alerts:
        type: integer
      total_farms:
        type: integer
      total_fields:
        type: integer
    type: object
...
  dto.HarvestResponse:
    properties:
      created_at:
        type: string
      created_by:
        type: integer
      field_id:
        type: integer
      field_loss:
        type: number
      field_name:
        type: string
      grain_moisture:
        type: number
      harvest_date:
        type: string
      id:
        type: integer
      impurity:
        type: number
      notes:
        type: string
      planting_id:
        type: integer
      productivity_bag_ha:
        type: number
      productivity_kg_ha:
        type: number
      total_bags:
        type: number
      updated_at:
        type: string
    type: object
...
  dto.MonitoringResponse:
    properties:
      created_at:
        type: string
      created_by:
        type: integer
      crop_stage:
        type: string
      field_id:
        type: integer
      field_name:
        type: string
      id:
        type: integer
      infestation_level:
        type: number
      inspection_date:
        type: string
      inspector:
        type: string
      notes:
        type: string
      planting_id:
        type: integer
      problem_name:
        type: string
      severity:
        $ref: '#/definitions/entities.Severity'
      technical_rec:
        type: string
      type:
        $ref: '#/definitions/entities.MonitoringType'
      updated_at:
        type: string
      urgent:
        type: boolean
    type: object
  dto.PaginatedResponse:
    properties:
      data: {}
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  dto.UpdateAlertStatusRequest:
    properties:
      status:
        allOf:
        - $ref: '#/definitions/entities.AlertStatus'
        enum:
        - open
        - resolved
        - ignored
    required:
    - status
    type: object
...
  entities.AlertPriority:
    enum:
    - low
    - medium
    - high
    type: string
    x-enum-varnames:
    - AlertPriorityLow
    - AlertPriorityMedium
    - AlertPriorityHigh
  entities.AlertStatus:
    enum:
    - open
    - resolved
    - ignored
    type: string
    x-enum-varnames:
    - AlertStatusOpen
    - AlertStatusResolved
    - AlertStatusIgnored
  entities.AlertType:
    enum:
    - low_stock
    - expired_input
    - pest
    - weather
    - pending_application
    - harvest_soon
    type: string
    x-enum-varnames:
    - AlertTypeLowStock
    - AlertTypeExpiredInput
    - AlertTypePest
    - AlertTypeWeather
    - AlertTypePending
    - AlertTypeHarvestSoon
...
  entities.MonitoringType:
    enum:
    - insect
    - disease
    - weed
    - nutritional
    type: string
    x-enum-comments:
      MonTypeDisease: doença
      MonTypeInsect: praga inseto
      MonTypeNutri: deficiência nutricional
      MonTypeWeed: planta daninha
    x-enum-descriptions:
    - praga inseto
    - doença
    - planta daninha
    - deficiência nutricional
    x-enum-varnames:
    - MonTypeInsect
    - MonTypeDisease
    - MonTypeWeed
    - MonTypeNutri
...
  entities.Severity:
    enum:
    - low
    - moderate
    - high
    - critical
    type: string
    x-enum-varnames:
    - SeverityLow
    - SeverityModerate
    - SeverityHigh
    - SeverityCritical
  utils.ErrorResponse:
    properties:
      code:
        type: string
      error:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API REST de gestão agrícola com autenticação JWT
  title: AgroControl API
  version: "1.0"
paths:
  /api/v1/alerts:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.AlertResponse'
            type: array
      security:
      - BearerAuth: []
      summary: Listar todos os alertas
      tags:
      - alerts
    post:
      consumes:
      - application/json
      description: Cria um alerta manual. Alertas automáticos são criados pelo sistema
        ao detectar estoque baixo ou monitoramento urgente.
      parameters:
      - description: Dados do alerta
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateAlertRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.AlertResponse'
      security:
      - BearerAuth: []
      summary: Criar alerta
      tags:
      - alerts
  /api/v1/alerts/{id}:
    get:
      parameters:
      - description: ID do alerta
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AlertResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Buscar alerta por ID
      tags:
      - alerts
  /api/v1/alerts/{id}/status:
    patch:
      consumes:
      - application/json
      description: Atualiza o status do alerta (open, resolved, ignored). Ao resolver,
        registra o timestamp de resolução.
      parameters:
      - description: ID do alerta
        in: path
        name: id
        required: true
        type: integer
      - description: Novo status
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateAlertStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AlertResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Atualizar status do alerta
      tags:
      - alerts
  /api/v1/alerts/open:
    get:
      description: Retorna apenas alertas com status 'open', ordenados por prioridade.
        Limite de 20 alertas.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.AlertResponse'
            type: array
      security:
      - BearerAuth: []
      summary: Listar alertas abertos
      tags:
      - alerts
...
  /api/v1/dashboard:
    get:
      description: 'Retorna métricas consolidadas: total de fazendas, talhões, plantios
        ativos, alertas abertos, insumos com estoque baixo e aplicações do mês.'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DashboardResponse'
      security:
      - BearerAuth: []
      summary: Dashboard consolidado
      tags:
      - dashboard
...
  /api/v1/fields/{id}/monitorings:
    get:
      parameters:
      - description: ID do talhão
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.MonitoringResponse'
            type: array
      security:
      - BearerAuth: []
      summary: Listar monitoramentos por talhão
      tags:
      - monitorings
  /api/v1/harvests:
    get:
      parameters:
      - description: Página
        in: query
        name: page
        type: integer
      - description: Itens por página
        in: query
        name: limit
        type: integer
      - description: Filtrar por talhão
        in: query
        name: field_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PaginatedResponse'
      security:
      - BearerAuth: []
      summary: Listar colheitas
      tags:
      - harvests
    post:
      consumes:
      - application/json
      description: Registra a colheita de um plantio ativo. Atualiza o status do plantio
        para 'harvested'. Valida que o plantio não foi colhido antes.
      parameters:
      - description: Dados da colheita
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateHarvestRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.HarvestResponse'
        "404":
          description: Plantio não encontrado
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "422":
          description: Plantio já colhido ou inativo
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Registrar colheita
      tags:
      - harvests
  /api/v1/harvests/{id}:
    get:
      parameters:
      - description: ID da colheita
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.HarvestResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Buscar colheita por ID
      tags:
      - harvests
...
  /api/v1/monitorings:
    get:
      description: Admin vê todos, outros veem apenas os seus
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.MonitoringResponse'
            type: array
      security:
      - BearerAuth: []
      summary: Listar monitoramentos
      tags:
      - monitorings
    post:
      consumes:
      - application/json
      description: Registra um monitoramento de pragas, doenças ou condições do talhão.
        Cria alerta automático se urgente ou severidade crítica.
      parameters:
      - description: Dados do monitoramento
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateMonitoringRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.MonitoringResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Registrar monitoramento
      tags:
      - monitorings
  /api/v1/monitorings/{id}:
    get:
      parameters:
      - description: ID do monitoramento
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MonitoringResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Buscar monitoramento por ID
      tags:
      - monitorings
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"